Add turbo to project (#11)
* 🎉 feat(turbo.json): add turbo.json configuration file for Turbo static site builder
The turbo.json file is added to the project root directory. It contains the configuration for the Turbo static site builder. The pipeline section defines two stages, build and lint. The build stage specifies the outputs to be included in the build, which are all files in the .next directory except for those in the .next/cache directory. The lint stage is currently empty.

* 🔥 chore(.gitignore): add .turbo, build/, dist/, and .next/ to ignore list
The .turbo directory and build/, dist/, and .next/ directories are now ignored by git. This is to prevent them from being accidentally committed to the repository.

* 🚀 chore(package.json): add turbo package to improve app performance
The turbo package has been added to the project to improve the performance of the application.