Update README.md with Turbo option (#19)
* 📝 docs(README.md): add instructions for using Turbo and troubleshooting
Added instructions for using Turbo to start the development server. Also added a note about a known issue with the turbod daemon process and how to run Turbo without it. This will help users who encounter the "Failed to Contact turbod" error.

* 🐛 fix(README.md): fix formatting of npm and turbo commands
The commands for starting the development server using NPM and Turbo were not properly formatted. The fix adds a space between the hyphen and the command to improve readability.

* Update README.md

Co-authored-by: Hector Sosa <hello@hectorsosa.me>

---------

Co-authored-by: Hector Sosa <hello@hectorsosa.me>