Introduce a Reset button (#12)
* 🚀 feat(Settings.tsx): add reset button to reset settings to default values
The reset button allows users to reset all settings to their default values. When the button is clicked, all settings are reset to their default values.

* 🔧 chore(Settings.tsx): refactor handleReset function to use defaultSettings object
The handleReset function has been refactored to use the defaultSettings object instead of hardcoding the default values. This improves maintainability and reduces the risk of errors when changing the default values. The value of the aspectRatio radio group is now set to the value of the settings.aspectRatio state variable to ensure that the correct value is selected when resetting the settings.

* 🐛 fix(defaults.ts): change default value of insetPadding from 0 to 1
The default value of insetPadding has been changed from 0 to 1 to provide a better visual experience. The previous value of 0 caused the image to be too close to the border, which made it difficult to see. The new value of 1 provides a small amount of padding, which makes the image more visible and easier to work with.

* 🎨 style(Button.tsx): re-order imports and add ghost variant
This commit reorders the imports to follow the convention of having external libraries first, then internal modules. It also adds a new ghost variant to the buttonVariants constant.

* 🎨 style(Settings.tsx): improve Settings UI layout and add reset button icon
The Settings UI layout has been improved by adding a title and a reset button icon. The title is now centered and the reset button is now an icon instead of text. This improves the overall look and feel of the Settings component.

* 🐛 fix(defaults.ts): change default value of insetPadding from 1 to 0
The default value of insetPadding has been changed from 1 to 0 to remove the unnecessary padding around the image.

* 🎨 style(Button.tsx): change ghost button style
The ghost button style has been updated to have a slate background color and white text color. The hover effect has also been updated to have a red background color with 90% opacity. This change improves the visibility of the ghost button and makes it more consistent with the design system.

* 🐛 fix(Settings.tsx): set default value for insetPadding to 2
The insetPadding value was not being reset to the default value when the reset button was clicked. This fix sets the default value to 2, which is the desired value.

* 🎨 style(Settings.tsx): remove hover effect and margin from reset button
The hover effect and margin on the reset button were removed to simplify the button's appearance and make it consistent with the other buttons in the component.